% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{getModel}
\alias{getModel}
\title{Retrieve a model description based on an identifier}
\usage{
getModel(id)
}
\arguments{
\item{id}{model identifier}
}
\description{
Retrieve a model description based on an identifier
}
