% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{getSearchResult}
\alias{getSearchResult}
\title{Fetch a list of models.}
\usage{
getSearchResult(query, offset = 0, all = FALSE)
}
\arguments{
\item{query}{Specify a search query. Can be omitted}

\item{offset}{Return results with an offset}

\item{all}{Should we make multiple queries to retrieve all results?}
}
\description{
Fetch a list of models.
}
\details{
This method can only return 10 results at a time.
This maxResults configured by the user preferences, rather than a query parameter.
See https://bitbucket.org/jummp/jummp/src/master/jummp-plugins/jummp-plugin-web-application/grails-app/controllers/net/biomodels/jummp/webapp/SearchController.groovy for more information.
}
