% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{print.ddmoreSearch}
\alias{print.ddmoreSearch}
\title{Print a ddmore search result in a nice format}
\usage{
\method{print}{ddmoreSearch}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a ddmore search result in a nice format
}
