% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{saveFiles}
\alias{saveFiles}
\title{Save the associated files of a model to disk}
\usage{
saveFiles(model, destination)
}
\arguments{
\item{model}{either the model identifier, or the downloaded model using `getModel`}

\item{destination}{destination directory}
}
\description{
Save the associated files of a model to disk
}
