% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{setBaseUrl}
\alias{setBaseUrl}
\title{Set the base URL to use as a repository
By default, we use the ddmore public repository}
\usage{
setBaseUrl(url)
}
\arguments{
\item{url}{The base url, e.g. http://repository.ddmore.eu}
}
\description{
Set the base URL to use as a repository
By default, we use the ddmore public repository
}
